+++ tests/questiontype_test.php
@@ -173,6 +173,7 @@ class qtype_opaque_test extends question_testcase {
         $expectedq->showhintafter = 0;
         $expectedq->showsolutionafter = 0;
         $expectedq->showsolutionaftertest = 0;
+        $q->numattemptlock = 0;
         $expectedq->exammode = 0;
 
         $this->assert(new question_check_specified_fields_expectation($expectedq), $q);
@@ -245,6 +246,7 @@ class qtype_opaque_test extends question_testcase {
         $expectedq->showhintafter = 0;
         $expectedq->showsolutionafter = 0;
         $expectedq->showsolutionaftertest = 0;
+        $q->numattemptlock = 0;
         $expectedq->exammode = 0;
 
         $this->assert(new question_check_specified_fields_expectation($expectedq), $q);